middleware/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
boto3==1.34.2
botocore==1.34.2
click==8.1.7
fastapi==0.105.0
h11==0.14.0
idna==3.6
jmespath==1.0.1
mysql-connector-python==8.2.0
protobuf==4.21.12
pydantic==2.5.2
pydantic_core==2.14.5
python-dateutil==2.8.2
s3transfer==0.9.0
shortuuid==1.0.11
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.9.0
urllib3==2.0.7
uvicorn==0.24.0.post1
